+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix detection if relevant part of LaTeX log for the case `! LaTeX Error: Something's wrong`
+
 ## [1.2.0] - 2023-09-25
 
 ### Added